+++ db/init.sql
@@ -12,9 +12,3 @@ CREATE TABLE IF NOT EXISTS invoice (
   isActive BOOLEAN NOT NULL DEFAULT true,
   deactiveAt DATETIME
 );
-
-INSERT INTO invoice
-  (id, document, description, amount, referenceMonth, referenceYear)
-VALUES
-  (UNHEX(REPLACE(UUID(), "-", "")), 'MC-Donalds', 'fast food', 100.24, NOW(), MONTH(NOW())),
-  (UNHEX(REPLACE(UUID(), "-", "")), 'C&A', 'general store', 1000222333.99, NOW(), MONTH(NOW()));
