+++ invoice_challenge/invoice_controller.py
@@ -3,15 +3,17 @@ from flask_restful import Resource
 from flask_restful.reqparse import RequestParser
 from werkzeug.exceptions import HTTPException
 from invoice_challenge.invoice_model import InvoiceModel
-import invoice_challenge.controller_helper as CH
+import invoice_challenge.helpers.controller_helper as CH
 
 # RequestParser serve as an allowlist params (old whitelist params).
-parser = RequestParser(bundle_errors=True)
-parser.add_argument("document",       type=str,   help="Please enter a valid String as a document")
-parser.add_argument("description",    type=str,   help="Please enter a valid String as a description")
-parser.add_argument("amount",         type=float, help="Please enter valid Decimal as an amount")
-parser.add_argument("referenceMonth", type=str,   help="Please enter a valid Datetime as a referenceMonth")
-parser.add_argument("referenceYear",  type=int,   help="Please enter valid Integer as a referenceYear")
+def invoice_request_parser():
+    parser = RequestParser(bundle_errors=True)
+    parser.add_argument("document",       type=str,   help="Please enter a valid String as a document")
+    parser.add_argument("description",    type=str,   help="Please enter a valid String as a description")
+    parser.add_argument("amount",         type=float, help="Please enter valid Decimal as an amount")
+    parser.add_argument("referenceMonth", type=str,   help="Please enter a valid Datetime as a referenceMonth")
+    parser.add_argument("referenceYear",  type=int,   help="Please enter valid Integer as a referenceYear")
+    return parser
 
 class Invoice(Resource):
     def get(self, id):
@@ -29,8 +31,8 @@ class Invoice(Resource):
 
     def put(self, id):
         try:
+            params = invoice_request_parser().parse_args()
             invoice_model = InvoiceModel()
-            params = parser.parse_args()
             invoice_model.update_item(id, params)
             return {}, 204
         except HTTPException as e:
@@ -57,8 +59,8 @@ class Invoice(Resource):
 class InvoiceCollection(Resource):
     def post(self):
         try:
+            params = invoice_request_parser().parse_args()
             invoice_model = InvoiceModel()
-            params = parser.parse_args()
             invoice_model.create_item(params)
             return {}, 201
         except HTTPException as e:
@@ -71,13 +73,14 @@ class InvoiceCollection(Resource):
 
     def get(self):
         try:
+            parser = invoice_request_parser()
             parser.add_argument("pageSize",       type=int, location="args")
             parser.add_argument("page",           type=int, location="args")
             parser.add_argument("order_by_desc",  type=str, location="args")
             parser.add_argument("order_by_asc",   type=str, location="args")
+            params = parser.parse_args()
 
             invoice_model = InvoiceModel()
-            params = parser.parse_args()
             data_resp = invoice_model.read_items(params)
 
             formmated_resp = CH.format_get_collection_response(data_resp, params)
